refactor: move IBMEventStreamsClient to event_streams module
Extracts IBMEventStreamsClient from gateway/core/ibm_cloud/clients.py into
its own gateway/core/ibm_cloud/event_streams/event_streams_client.py module,
following the same pattern as the cos/ and code_engine/ submodules. Updates
all imports and moves the corresponding tests to a matching test subpackage.
